requirements.txt
scapy.all
geoip2.database

